MINITALK FINI(talk)

+++ srcs/client.c
@@ -6,13 +6,13 @@
 /*   By: vsozonof <vsozonof@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/14 13:40:04 by vsozonof          #+#    #+#             */
-/*   Updated: 2023/02/27 01:09:54 by vsozonof         ###   ########.fr       */
+/*   Updated: 2023/02/28 08:32:30 by vsozonof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minitalk.h"
 
-sig_atomic_t	g_flag = 0;
+sig_atomic_t	g_sig_client = 0;
 
 void	ft_client_sig_handler(int sig, siginfo_t *sig_info, void *context)
 {
@@ -20,44 +20,42 @@ void	ft_client_sig_handler(int sig, siginfo_t *sig_info, void *context)
 	(void)sig_info;
 	(void)context;
 	if (sig == SIGUSR1)
+		g_sig_client = 1;
+	else if (sig == SIGUSR2)
 	{
-		g_flag = 1;
-		ft_printf("sig recu\n");
+		ft_printf("The string was fully passed to the server.");
+		exit(EXIT_SUCCESS);
 	}
 }
 
 void	ft_char_handler(int pid, unsigned char c)
 {
-	int	value;
+	int				bit_index;
+	unsigned char	bit_cmp;
 
-	value = c;
-	while (value >= 0)
+	bit_index = 7;
+	bit_cmp = 1u << bit_index;
+	while (bit_index >= 0)
 	{
-		g_flag = 0;
-		if (value > 0)
-		{
+		g_sig_client = 0;
+		if (bit_cmp & c)
 			kill(pid, SIGUSR1);
-		}
-		if (value == 0)
-			return ;
-		value--;
-		usleep(2000);
-		if (g_flag == 0)
+		else
+			kill(pid, SIGUSR2);
+		while (!g_sig_client)
 			pause();
+		bit_index--;
+		bit_cmp >>= 1;
 	}
 }
 
 void	ft_str_handler(int pid, char *str)
-{
+{	
 	int	i;
 
 	i = -1;
-	ft_printf("%i - %s\n", pid, str);
 	while (str[++i])
-	{
 		ft_char_handler(pid, str[i]);
-		kill(pid, SIGUSR2);
-	}
 	ft_char_handler(pid, 0);
 }
 
@@ -71,6 +69,8 @@ int	ft_arg_checker(int argc, char **argv)
 		return (ft_error_handler("Error: invalid input."));
 	else if (ft_overflow_handler(argv[1]) != 1)
 		return (ft_error_handler("Error: invalid PID."));
+	else if (ft_atol(argv[1]) < 10)
+		return (ft_error_handler("Error: invalid PID"));
 	return (1);
 }
 
@@ -83,7 +83,7 @@ int	main(int argc, char **argv)
 	sa.sa_flags = SA_SIGINFO;
 	sigemptyset(&sa.sa_mask);
 	sigaction(SIGUSR2, &sa, NULL);
-    sigaction(SIGUSR1, &sa, NULL);
+	sigaction(SIGUSR1, &sa, NULL);
 	if (ft_arg_checker(argc, argv) != 1)
 		return (0);
 	pid = ft_atol(argv[1]);
